+++ src/pages/dashboard.js
@@ -1,6 +1,5 @@
 import { getError as FoxyError } from '../util/foxyErrors';
 import Alert from '../util/alerts';
-import * as FoxyStorage from '../util/FoxyStorage';
 import * as MangaTable from './util/manga-table';
 import * as Import from './util/import';
 import Export from './util/export';
@@ -11,7 +10,9 @@ import Export from './util/export';
  */
 async function initTable(table) {
   try {
-    const bookmarkList = await FoxyStorage.getMetadata();
+    const bookmarkList = await browser.runtime.sendMessage({
+      type: 'get-bookmark-data',
+    });
 
     const promises = [];
     bookmarkList.forEach(async (bookmark) => {